Issue #171 - Fixing Lint Warnings in the Explore Packages
Fixing Some Android Lint and SonarLint Warnings in these packages:
* `fr.free.nrw.commons.explore`
* `fr.free.nrw.commons.explore.categories`
* `fr.free.nrw.commons.explore.images`
* `fr.free.nrw.commons.explore.recentsearches`

Some of the fixes were:
* Switching the try/catch/finally blocks for database querying in `RecentSearchesDao` into try-with-resources blocks, with the Cursor object getting set as the closable resource
* Switching one field type in the sqlite "Create Table Statement" in `RecentSearchesDao` from STRING to TEXT
* Creating a `DAOException` and throwing that instead of a generic RuntimeException in `RecentSearchesDao`
* Logging some Exceptions using Timber instead of `e.printStackTrace()`
* A few access modifier changes to be more restrictive
* Adding a private constructor to some static classes
* Made some constant String variables for recurring Strings
* Changed the switch cases with only one case into if/else statements
* Added some `@NonNull` and `@Nullable` annotations to methods that inherit methods that use those annotations
* Added `final` to some constants

+++ app/src/main/java/fr/free/nrw/commons/explore/recentsearches/RecentSearchesContentProvider.java
@@ -32,6 +32,7 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
     private static final int RECENT_SEARCHES = 1;
     private static final int RECENT_SEARCHES_ID = 2;
     private static final String BASE_PATH = "recent_searches";
+    private static final String UNKNOWN_URL = "Unknown URI: ";
     public static final Uri BASE_URI = Uri.parse("content://" + BuildConfig.RECENT_SEARCH_AUTHORITY + "/" + BASE_PATH);
     private static final UriMatcher uriMatcher = new UriMatcher(NO_MATCH);
 
@@ -77,7 +78,7 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
                 );
                 break;
             default:
-                throw new IllegalArgumentException("Unknown URI" + uri);
+                throw new IllegalArgumentException(UNKNOWN_URL + uri);
         }
 
         cursor.setNotificationUri(getContext().getContentResolver(), uri);
@@ -99,12 +100,10 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
         int uriType = uriMatcher.match(uri);
         SQLiteDatabase sqlDB = dbOpenHelper.getWritableDatabase();
         long id;
-        switch (uriType) {
-            case RECENT_SEARCHES:
-                id = sqlDB.insert(TABLE_NAME, null, contentValues);
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown URI: " + uri);
+        if (uriType == RECENT_SEARCHES) {
+            id = sqlDB.insert(TABLE_NAME, null, contentValues);
+        } else {
+            throw new IllegalArgumentException(UNKNOWN_URL + uri);
         }
         getContext().getContentResolver().notifyChange(uri, null);
         return Uri.parse(BASE_URI + "/" + id);
@@ -118,16 +117,14 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
         int rows;
         int uriType = uriMatcher.match(uri);
         SQLiteDatabase db = dbOpenHelper.getReadableDatabase();
-        switch (uriType) {
-            case RECENT_SEARCHES_ID:
-                Timber.d("Deleting recent searches id %s", uri.getLastPathSegment());
-                rows = db.delete(RecentSearchesDao.Table.TABLE_NAME,
-                        "_id = ?",
-                        new String[]{uri.getLastPathSegment()}
-                );
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown URI" + uri);
+        if (uriType == RECENT_SEARCHES_ID) {
+            Timber.d("Deleting recent searches id %s", uri.getLastPathSegment());
+            rows = db.delete(RecentSearchesDao.Table.TABLE_NAME,
+                    "_id = ?",
+                    new String[]{uri.getLastPathSegment()}
+            );
+        } else {
+            throw new IllegalArgumentException(UNKNOWN_URL + uri);
         }
         getContext().getContentResolver().notifyChange(uri, null);
         return rows;
@@ -143,15 +140,13 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
         int uriType = uriMatcher.match(uri);
         SQLiteDatabase sqlDB = dbOpenHelper.getWritableDatabase();
         sqlDB.beginTransaction();
-        switch (uriType) {
-            case RECENT_SEARCHES:
-                for (ContentValues value : values) {
-                    Timber.d("Inserting! %s", value);
-                    sqlDB.insert(TABLE_NAME, null, value);
-                }
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown URI: " + uri);
+        if (uriType == RECENT_SEARCHES) {
+            for (ContentValues value : values) {
+                Timber.d("Inserting! %s", value);
+                sqlDB.insert(TABLE_NAME, null, value);
+            }
+        } else {
+            throw new IllegalArgumentException(UNKNOWN_URL + uri);
         }
         sqlDB.setTransactionSuccessful();
         sqlDB.endTransaction();
@@ -178,21 +173,19 @@ public class RecentSearchesContentProvider extends CommonsDaggerContentProvider 
         int uriType = uriMatcher.match(uri);
         SQLiteDatabase sqlDB = dbOpenHelper.getWritableDatabase();
         int rowsUpdated;
-        switch (uriType) {
-            case RECENT_SEARCHES_ID:
-                if (TextUtils.isEmpty(selection)) {
-                    int id = Integer.valueOf(uri.getLastPathSegment());
-                    rowsUpdated = sqlDB.update(TABLE_NAME,
-                            contentValues,
-                            COLUMN_ID + " = ?",
-                            new String[]{String.valueOf(id)});
-                } else {
-                    throw new IllegalArgumentException(
-                            "Parameter `selection` should be empty when updating an ID");
-                }
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown URI: " + uri + " with type " + uriType);
+        if (uriType == RECENT_SEARCHES_ID) {
+            if (TextUtils.isEmpty(selection)) {
+                int id = Integer.valueOf(uri.getLastPathSegment());
+                rowsUpdated = sqlDB.update(TABLE_NAME,
+                        contentValues,
+                        COLUMN_ID + " = ?",
+                        new String[]{String.valueOf(id)});
+            } else {
+                throw new IllegalArgumentException(
+                        "Parameter `selection` should be empty when updating an ID");
+            }
+        } else {
+            throw new IllegalArgumentException(UNKNOWN_URL + uri + " with type " + uriType);
         }
         getContext().getContentResolver().notifyChange(uri, null);
         return rowsUpdated;
